Fixes, small game and a beta project (v.0.4.1)
The Calculator sub has been given a refresher, TicTacs should be ignored ;3 the Python3 in a sandwich is a WIP project of summing up various things related to Python and Python 3. Might not look good as I've been using a custom coloring scheme in Atom. A redo might happen once this is finished (if it ever will be.)